+++ final-capstone-project/src/main/java/com/techelevator/model/QuestionDAO.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 public interface QuestionDAO {
 	
-	public List<Question> getQuestionsBySurveyId(int id); 
+	public List<Question> getQuestionsBySurveyId(long id); 
 
 }
